объедитение двух модулей в big-picture.js, удаление rendering-big-picture.js
